fix change uint64_t into int

+++ lab1/main.cpp
@@ -5,7 +5,7 @@ int main()
 {
     int start, end;
     std::cin >> start >> end;
-    uint64_t ans = how_one_in_cons(start, end);
+    int ans = how_one_in_cons(start, end);
     std::cout << ans << std::endl;
     return 0;
 }
